Q4-2021: Force charts javascript version
You get issues that it caches old version of the charts file when for example using the preview site and when it goes live you don't have the correct latest version.

This should force it to not be cached for users which is helpful for future fixes or additions to older progress reports.