DownloadButton: Fix Installer and Portable Windows (#258)
Stable for windows will get Installer or Portable if it's in the assetname or just Download, Linux will get Flatpak or AppImage in the name or just the original displayname and MacOS will get Download just like Windows. And every display will be capitalised like binary for Linux will now be Binary.